Move DragDropPayload to Classes
